Adds default(present) to acc parallel loops
After using this code for some tutorials a common issue I saw was
students hitting an error due to a variable being unknown. I tracked
this down to the fact that the code was written to be build with the
`defpresent` option (or managed works too). This change makes the code
more portable by adding the default(present) to the loops explicitly.